+++ app/gui/controller/double-representation/src/text.rs
@@ -257,8 +257,8 @@ mod test {
         fn assert_same_node_ids(&self, ast1: &ast::known::Module, ast2: &ast::known::Module) {
             let ids1 = main_nodes(ast1);
             let ids2 = main_nodes(ast2);
-            DEBUG!("IDs1: {ids1:?}");
-            DEBUG!("IDs2: {ids2:?}");
+            debug!("IDs1: {ids1:?}");
+            debug!("IDs2: {ids2:?}");
             assert_eq!(ids1, ids2, "Node ids mismatch in {:?}", self);
         }
     }