+++ app/gui/view/graph-editor/src/component/node/output/area.rs
@@ -352,8 +352,14 @@ impl Model {
             if DEBUG {
                 let indent  = " ".repeat(4*builder.depth);
                 let skipped = if !is_a_port { "(skip)" } else { "" };
-                DEBUG!("{indent}[{node.payload.index},{node.payload.length}] \
-                {skipped} {node.kind.variant_name():?} (tp: {node.tp():?}) (id: {node.ast_id:?})");
+                debug!(
+                    "{indent}[{},{}] {skipped} {:?} (tp: {:?}) (id: {:?})",
+                    node.payload.index,
+                    node.payload.length,
+                    node.kind.variant_name(),
+                    node.tp(),
+                    node.ast_id
+                );
             }
 
             if is_a_port {
@@ -412,7 +418,7 @@ impl Model {
     fn set_expression(&self, new_expression: impl Into<node::Expression>) {
         let new_expression = Expression::from(new_expression.into());
         if DEBUG {
-            DEBUG!("\n\n=====================\nSET EXPR: {new_expression:?}")
+            debug!("\n\n=====================\nSET EXPR: {new_expression:?}")
         }
 
         self.set_label_on_new_expression(&new_expression);