+++ app/gui/view/src/debug_mode_popup.rs
@@ -104,19 +104,17 @@ impl PopupLabel {
 struct Model {
     display_object: display::object::Instance,
     label:          PopupLabel,
-    logger:         Logger,
 }
 
 impl Model {
     /// Constructor.
     pub fn new(app: &Application) -> Self {
-        let logger = Logger::new("DebugModePopup");
         let display_object = display::object::Instance::new();
         let label = PopupLabel::new(app);
         label.set_delay(LABEL_VISIBILITY_DELAY_MS);
         display_object.add_child(&label);
 
-        Self { display_object, label, logger }
+        Self { display_object, label }
     }
 
     /// Show "Debug Mode enabled" label.